requirements.txt
attrs==17.4.0
nose==1.3.7
pluggy==0.6.0
py==1.5.2
pybind11==2.2.2
pyfrc==2018.2.1
pynetconsole==2.0.2
pynetworktables==2018.1.0
pytest==3.4.2
robotpy-ctre==2018.2.5
robotpy-hal-base==2018.1.6
robotpy-hal-sim==2018.1.6
robotpy-installer==2018.0.2
robotpy-wpilib-utilities==2018.1.4
six==1.11.0
wpilib==2018.1.6
